[interjections] - move remove interjection to it's own class
